+++ js/pools.js
@@ -1,0 +1,12 @@
+export const titlesPool = [
+  'Hilton','Kempinski','Crown Plaza','Holiday Inn','Four Seasons','Kamikadze','Sushi Hotel','Kawasaki','Sudoku','Sumo'
+];
+export const typesPool = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+export const checkInPool = ['12:00', '13:00', '14:00'];
+export const featuresPool = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+export const descriptionsList = ['sea view','park view','infinity pool','king size bed', 'animals allowed', 'not far from city center', 'bed for baby', 'smoking allowed'];
+export const photosPool = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+];
